+++ pykeepass/__init__.py
@@ -1,2 +1,4 @@
 from __future__ import absolute_import
-from .pykeepass import PyKeePass
+from .pykeepass import PyKeePass, create_database
+
+from pykeepass.version import __version__