+++ pykeepass/kdbx_parsing/kdbx3.py
@@ -1,11 +1,10 @@
-#!/bin/env python3
 # Evan Widloski - 2018-04-11
 # keepass decrypt experimentation
 
 import hashlib
 from construct import (
-    Byte, Bytes, Int16ul, Int32ul, RepeatUntil, GreedyBytes, Struct, this,
-    Mapping, Switch, Prefixed, Padding, Checksum, Computed, IfThenElse,
+    Byte, Bytes, Int16ul, Int32ul, Int64ul, RepeatUntil, GreedyBytes, Struct,
+    this, Mapping, Switch, Prefixed, Padding, Checksum, Computed, IfThenElse,
     Pointer, Tell, len_
 )
 from .common import (
@@ -67,7 +66,7 @@ DynamicHeaderItem = Struct(
             this.id,
             {'compression_flags': CompressionFlags,
              'cipher_id': CipherId,
-             'transform_rounds': Int32ul,
+             'transform_rounds': Int64ul,
              'protected_stream_id': ProtectedStreamId
              },
             default=GreedyBytes
@@ -129,7 +128,7 @@ PayloadBlocks = RepeatUntil(
 UnpackedPayload = Reparsed(
     Struct(
         # validate payload decryption
-        Checksum(
+        "cred_check" / Checksum(
             Bytes(32),
             lambda this: this._._.header.value.dynamic_header.stream_start_bytes.data,
             this,